+++ CHANGELOG.md
@@ -24,7 +24,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/report/compare/release-3.14.0...master
+[Unreleased]: https://github.com/atlassian/report/compare/release-3.15.0...master
+
+## [3.15.0] - 2023-07-03
+[3.15.0]: https://github.com/atlassian/report/compare/release-3.14.0...release-3.15.0
 
 ### Added
 - Add `LatencyImpact` and `LatencyImpact.Builder`. Unblock [JPERF-1160].